add put to order route

+++ routes/order.routes.js
@@ -14,7 +14,7 @@ const { isAuthenticated } = require("../middleware/jwt.middleware")
 router.post("/orders", isAuthenticated, async (req, res, next) => {
 
   try {
-    const { restaurantId } = req.body
+    const { restaurantId, products, total } = req.body
 
     if (!mongoose.Types.ObjectId.isValid(restaurantId)) {
       return res.status(400).json({ message: "Invalid restaurant id" })
@@ -29,8 +29,8 @@ router.post("/orders", isAuthenticated, async (req, res, next) => {
     const createdOrder = await Order.create({
       user: req.payload._id,
       restaurant: restaurantId,
-      products: [],
-      total: 0
+      products: products,
+      total: total
     })
 
     res.status(201).json(createdOrder)
@@ -44,65 +44,88 @@ router.post("/orders", isAuthenticated, async (req, res, next) => {
 
 // PUT /orders/:orderId
 
-router.put("/orders/:orderId", isAuthenticated, async (req, res, next) => {
+// router.put("/orders/:orderId", isAuthenticated, async (req, res, next) => {
+
+//   try {
+//     const { orderId } = req.params
+//     const { productId, quantity } = req.body
+
+//     if (
+//       !mongoose.Types.ObjectId.isValid(orderId) ||
+//       !mongoose.Types.ObjectId.isValid(productId)
+//     ) {
+//       return res.status(400).json({ message: "Invalid id" })
+//     }
+
+//     const order = await Order.findById(orderId)
+
+//     if (!order) {
+//       return res.status(404).json({ message: "Order not found" })
+//     }
+
+//     if (order.user.toString() !== req.payload._id) {
+//       return res.status(403).json({ message: "Unauthorized" })
+//     }
+
+//     const product = await Product.findById(productId)
+
+//     if (!product) {
+//       return res.status(404).json({ message: "Product not found" })
+//     }
+
+//     // VALIDACIÓN CLAVE
+//     if (product.restaurant.toString() !== order.restaurant.toString()) {
+//       return res.status(400).json({
+//         message: "Product does not belong to this restaurant"
+//       })
+//     }
+
+//     // Buscar si ya existe en products
+//     const existingItem = order.products.find(item =>
+//       item.product.toString() === productId
+//     )
+
+//     if (existingItem) {
+//       existingItem.quantity += quantity || 1
+//     } else {
+//       order.products.push({
+//         product: productId,
+//         quantity: quantity || 1,
+//         price: product.price
+//       })
+//     }
+
+//     // recalcular total
+//     order.total = order.products.reduce((acc, item) => {
+//       return acc + item.price * item.quantity
+//     }, 0)
+
+//     await order.save()
+
+//     res.json(order)
+
+//   } catch (err) {
+//     console.error("Error updating a order", err)
+//     next(err)
+//   }
+// })
+
+// PUT /orders/:orderId/status
+
+router.put("/orders/:orderId/status", isAuthenticated, async (req, res, next) => {
 
   try {
+
     const { orderId } = req.params
-    const { productId, quantity } = req.body
+    const { status } = req.body
 
-    if (
-      !mongoose.Types.ObjectId.isValid(orderId) ||
-      !mongoose.Types.ObjectId.isValid(productId)
-    ) {
-      return res.status(400).json({ message: "Invalid id" })
-    }
-
-    const order = await Order.findById(orderId)
-
-    if (!order) {
-      return res.status(404).json({ message: "Order not found" })
-    }
-
-    if (order.user.toString() !== req.payload._id) {
-      return res.status(403).json({ message: "Unauthorized" })
-    }
-
-    const product = await Product.findById(productId)
-
-    if (!product) {
-      return res.status(404).json({ message: "Product not found" })
-    }
-
-    // VALIDACIÓN CLAVE
-    if (product.restaurant.toString() !== order.restaurant.toString()) {
-      return res.status(400).json({
-        message: "Product does not belong to this restaurant"
-      })
-    }
-
-    // Buscar si ya existe en products
-    const existingItem = order.products.find(item =>
-      item.product.toString() === productId
+    const updateOrder = await Order.findByIdAndUpdate(
+      orderId,
+      { status },
+      { new: true }
     )
 
-    if (existingItem) {
-      existingItem.quantity += quantity || 1
-    } else {
-      order.products.push({
-        product: productId,
-        quantity: quantity || 1,
-        price: product.price
-      })
-    }
-
-    // recalcular total
-    order.total = order.products.reduce((acc, item) => {
-      return acc + item.price * item.quantity
-    }, 0)
-
-    await order.save()
-
-    res.json(order)
+    res.json(updateOrder)
 
   } catch (err) {
     console.error("Error updating a order", err)
@@ -162,39 +185,6 @@ router.get("/orders/:orderId", isAuthenticated, async (req, res, next) => {
 })
 
 
-// DELETE /orders/:orderId/products/:productId
-
-router.delete("/orders/:orderId/products/:productId", isAuthenticated, async (req, res, next) => {
-
-  try {
-    const { orderId, productId } = req.params
-
-    const order = await Order.findById(orderId)
-
-    if (!order) {
-      return res.status(404).json({ message: "Order not found" })
-    }
-
-    if (order.user.toString() !== req.payload._id) {
-      return res.status(403).json({ message: "Unauthorized" })
-    }
-
-    order.products = order.products.filter(item =>
-      item.product.toString() !== productId
-    )
-
-    order.total = order.products.reduce((acc, item) => {
-      return acc + item.price * item.quantity
-    }, 0)
-
-    await order.save()
-
-    res.json(order)
-
-  } catch (err) {
-    next(err)
-  }
-})
 
 // DELETE /orders/:orderId
 
@@ -205,6 +195,16 @@ router.delete("/orders/:orderId", isAuthenticated, async (req, res, next) => {
 
     if (!mongoose.Types.ObjectId.isValid(orderId)) {
       return res.status(400).json({ message: "Invalid id" })
+    }
+
+    const order = await Order.findById(orderId)
+
+    if (!order) {
+      return res.status(404).json({ message: "Order not found" })
+    }
+
+    if (order.user.toString() !== req.payload._id) {
+      return res.status(403).json({ message: "Unauthorized" })
     }
 
     const deletedOrder = await Order.findByIdAndDelete(orderId)
